ISSARA_CODE_CLEANED_OF_DATA/ilm-server-2/requirements.txt
Django>=2.0,<3.0
djangorestframework>=3.0,<4.0
psycopg2-binary
pymysql
mysqlclient
PyJWT==1.7.1
djangorestframework_simplejwt==4.5.0
django-cors-headers
django-debug-toolbar==2.1
django-environ
slacker-log-handler
drf-renderer-xlsx==0.3.7
simplejson
djangosecure
django-sslserver
django-rest-passwordreset
phonenumbers
drf-jwt-2fa==0.2.0
factory-boy
django-storages[google]
tblib==1.7.0
